onboarding/pyproject.toml
[tool.poetry]
name = "onboarding"
version = "0.1.0"
description = ""
authors = ["paulburkhardt <92929336+paulburkhardt@users.noreply.github.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
